perf(WebAPI): Minor performance improvement how strings are converted from and to enums
