+++ src/Domain/Extensions/EnumMappers/EnumMapperExtensions.Base.cs
@@ -1,0 +1,8 @@
+using Logging.Interface;
+
+namespace PlexRipper.Domain;
+
+public static partial class EnumMapperExtensions
+{
+    private static readonly ILog _log = LogManager.CreateLogInstance(typeof(Domain.EnumMapperExtensions));
+}
